requirements_test.txt
pyelasticsearch
pysolr
django-nose
mock
nose
coverage
coveralls
dj-database-url==0.4.2
